First steps of splitting distributor logic up into disjoint stripes
For now only a single stripe may be used.
